% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw2d.R
\name{draw2d}
\alias{draw2d}
\title{plot ride track data on a map}
\usage{
draw2d(geodf, outfile, maptitle, definedmaps, usemap, maptype = "osm",
  minTiles = 50, mapsize = c(1600, 1200), fine.map = FALSE,
  margin.factor = 0.08, draw.speed = FALSE, line.color = "magenta",
  line.width = 3, line.alpha = 0.8, speed.color = "speedcolors",
  speed.alpha = 0.7, speed.ptsize = 6, speed.pch = 19,
  jpeg.quality = 90)
}
\arguments{
\item{geodf}{data frame or tibble containing at least: position_lat.dd,
position_lon.dd,(or lat,lon)(both numeric,decimal degrees),
startbutton.date(int),startbutton.time(int),segment(numeric),
and speed.m.s for speed coloring}

\item{outfile}{name of output file, extension is either .tiff or .jpg}

\item{maptitle}{string containing title}

\item{definedmaps}{list, each named entry is a list of two vectors of
length two, named lat and lon, containing the min and max of the
latitude andf longitude of the map area}

\item{usemap}{string identifying area from list \eqn{definedmaps} or}

\item{maptype}{map to use as background (\code{"maptoolkit-topo"} or
\code{"bing"} or \code{"osm"} or \code{"stamen-terrain"} or
\code{"esri-topo"}  or \code{"stamen-watercolor"} or \code{"mapbox"} or
\code{"esri"} or \code{"osm-public-transport"} or \code{"opencyclemap"}
or \code{"apple-iphoto"} or \code{"skobbler"})}

\item{minTiles}{minimum number of tiles fetched for map, larger is slower
but better quality}

\item{mapsize}{pixel size of map created}

\item{fine.map}{if true, use c(7680,4800) for map size}

\item{margin.factor}{percentage to expand map area beyond limits of
the track in each direction, when autodetermining region to plot}

\item{draw.speed}{if true, draw track(s) as a series of points whose color
indicates travel speed}

\item{line.color}{the color to draw the lines of the tracks,
if a palette is specified (\code{"plasma"} or \code{"viridis"} or
 \code{"rainbow"} or \code{"heat"} or \code{"red-blue"})is specified,
each track supplied will be assigned a color from that palette}

\item{line.width}{the width of the line for the tracks}

\item{line.alpha}{the opacity of the line}

\item{speed.color}{palette to use to represent speed on plot
(\code{"speedcolors"} or \code{"red-blue-green"} or
\code{"rainbow"} or \code{"plasma"} or \code{"magma"} or
\code{"heat"} )}

\item{speed.alpha}{opacity of the speed line if \eqn{speed.color} specified}

\item{speed.ptsize}{size for the symbols used to plot the speed line,
doubled if \eqn{fine.map} is set true.}

\item{speed.pch}{the character to use plotting the points of the speed line}

\item{jpeg.quality}{the "quality" of the JPEG image, a percentage. Smaller
values will give more compression but also more degradation of the image}
}
\value{
NULL or plotly object
}
\description{
\code{draw2d} Plot lat/long data on maps, optionally using colored
  dots to indicate speed
}
\details{
Create a map displaying the tracks specified.  Tracks may be drawn all in one
  specified color, with each separate track drawn in a different color from
  a specified palette, or as a series of closely spaced dots (which may
  be prodded into appearing as a line through judicious choice of point
  size and alpha) with the color varying based on speed and the chosen
  palette.
}
