% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptrack2d.R
\name{maptrack2d}
\alias{maptrack2d}
\title{maptrack2d: Draw 2-d map displaying tracks}
\description{
This package also allows for the creation of maps of a ride or set of rides.
 The map area will be slightly larger than the track of the ride, unless you supply
 a named list of map lat/long boundaries. If a specific map is not specified from
 the list, it will choose the first in the list that covers the track.  The
 map is created using the OpenStreetMap package, and there are many choices for
 the map format.  Respect the terms of use and do not automate the generation
 of maps.
}
\section{main call}{

  \link{draw2d}
}

